style: separete public items routes
