+++ backend/kemomimi-api/src/routes.rs
@@ -1,0 +1,1 @@
+pub mod public_items;
