+++ backend/kemomimi-api/src/routes/public_items.rs
@@ -1,0 +1,1 @@
+pub mod index;
